1.0.4:修改打包脚本。增加CentOS 6服务脚本。

+++ mudis/src/pidfile.cpp
@@ -12,9 +12,16 @@ namespace lyramilk{ namespace proc
 		}
 	}
 
+	void pidfile::detach()
+	{
+		this->pf.clear();
+	}
+
 	pidfile::~pidfile()
 	{
-		unlink(this->pf.c_str());
+		if(!this->pf.empty()){
+			unlink(this->pf.c_str());
+		}
 	}
